Replaces a lot of N+1 queries for checking permissions.

It might make checking permissions for a single repo a bit slower, but not much.

With a local dataset I have measured the following changes on:
* /api/v1/repos/search?limit=1 => about same speed,maybe 0.9 speed (old time: 0,07s)
* /api/v1/repos/search?limit=50 => 2.6 speed increase (old time: 0.35s)
* /api/v1/repos/search?limit=500 => 3.8 speed increase (old time: 2.7s)
* /api/v1/repos/search?limit=1000 => 4.1 speed increase (old time: 5.4s)